+++ SampleScripts/dri/driDeployTanzuCluster.ps1
@@ -5,11 +5,16 @@
 
 <#
     .NOTES
-    ===========================================================================
+    ===================================================================================================================
     Created by:  Gary Blake - Senior Staff Solutions Architect
     Date:   2021-03-23
     Copyright 2021-2022 VMware, Inc.
-    ===========================================================================
+    ===================================================================================================================
+    .CHANGE_LOG
+
+    - 1.1.000   (Gary Blake / 2022-10-03) - Added support for VCF 4.5.x Planning and Prep Workbook
+
+    ===================================================================================================================
     
     .SYNOPSIS
     Deploy a Tanzu Cluster for Developer Ready Infrastructure
@@ -71,7 +76,7 @@ Try {
         $pnpWorkbook = Open-ExcelPackage -Path $Workbook
 
         Write-LogMessage -type INFO -message "Checking Valid Planning and Prepatation Workbook Provided"
-        if (($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.3.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.4.x")) {
+        if (($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.3.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.4.x") -and ($pnpWorkbook.Workbook.Names["vcf_version"].Value -ne "v4.5.x")) {
             Write-LogMessage -type INFO -message "Planning and Prepatation Workbook Provided Not Supported" -colour Red 
             Break
         }
@@ -104,19 +109,19 @@ Try {
     if ( $StatusMsg ) { Write-LogMessage -Type INFO -Message "$StatusMsg"; $ErrorMsg = $null } if ( $WarnMsg ) { Write-LogMessage -Type WARNING -Message $WarnMsg -Colour Magenta } if ( $ErrorMsg ) { Write-LogMessage -Type ERROR -Message $ErrorMsg -Colour Red }
 
     # Assign the New Tanzu Cluster Namespace Roles to Active Directory Groups
-    Write-LogMessage -Type INFO -Message "Assign the New Tanzu Cluster Namespace Roles to Active Directory Groups"
+    Write-LogMessage -Type INFO -Message "Attempting to Assign the New Tanzu Cluster Namespace Roles to Active Directory Groups"
     $StatusMsg = Add-NamespacePermission -server $sddcManagerFqdn -user $sddcManagerUser -pass $sddcManagerPass -sddcDomain $wldSddcDomainName -domain $domainFqdn -domainBindUser $domainBindUser -domainBindPass $domainBindPass -namespace $wmNamespaceName -principal $wmNamespaceEditUserGroup -role edit -type group -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -WarningVariable WarnMsg -ErrorVariable ErrorMsg
     if ( $StatusMsg ) { Write-LogMessage -Type INFO -Message "$StatusMsg" } if ( $WarnMsg ) { Write-LogMessage -Type WARNING -Message $WarnMsg -Colour Magenta } if ( $ErrorMsg ) { Write-LogMessage -Type ERROR -Message $ErrorMsg -Colour Red }
     $StatusMsg = Add-NamespacePermission -server $sddcManagerFqdn -user $sddcManagerUser -pass $sddcManagerPass -sddcDomain $wldSddcDomainName -domain $domainFqdn -domainBindUser $domainBindUser -domainBindPass $domainBindPass -namespace $wmNamespaceName -principal $wmNamespaceViewUserGroup -role view -type group -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -WarningVariable WarnMsg -ErrorVariable ErrorMsg
     if ( $StatusMsg ) { Write-LogMessage -Type INFO -Message "$StatusMsg" } if ( $WarnMsg ) { Write-LogMessage -Type WARNING -Message $WarnMsg -Colour Magenta } if ( $ErrorMsg ) { Write-LogMessage -Type ERROR -Message $ErrorMsg -Colour Red }
 
     # Enable a Virtual Machine Class for the Tanzu Kubernetes Cluster
-    Write-LogMessage -Type INFO -Message "Enable a Virtual Machine Class for the Tanzu Kubernetes Cluster"
+    Write-LogMessage -Type INFO -Message "Attempting to Enable a Virtual Machine Class for the Tanzu Kubernetes Cluster"
     $StatusMsg = Add-NamespaceVmClass -server $sddcManagerFqdn -user $sddcManagerUser -pass $sddcManagerPass -domain $wldSddcDomainName -Namespace $wmNamespaceName -VMClass $vmClass -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -WarningVariable WarnMsg -ErrorVariable ErrorMsg
     if ( $StatusMsg ) { Write-LogMessage -Type INFO -Message "$StatusMsg" } if ( $WarnMsg ) { Write-LogMessage -Type WARNING -Message $WarnMsg -Colour Magenta } if ( $ErrorMsg ) { Write-LogMessage -Type ERROR -Message $ErrorMsg -Colour Red }
 
     # Provision a Tanzu Kubernetes Cluster
-    Write-LogMessage -Type INFO -Message "Provision a Tanzu Kubernetes Cluster"
+    Write-LogMessage -Type INFO -Message "Attempting to Provision a Tanzu Kubernetes Cluster"
     $StatusMsg = Add-TanzuKubernetesCluster -server $sddcManagerFqdn -user $sddcManagerUser -pass $sddcManagerPass -domain $wldSddcDomainName -cluster $wmClusterName -yaml .\SampleYaml\sfo-w01-tkc01-cluster.yaml -WarningAction SilentlyContinue -ErrorAction SilentlyContinue -WarningVariable WarnMsg -ErrorVariable ErrorMsg
     if ( $StatusMsg ) { Write-LogMessage -Type INFO -Message "$StatusMsg" } if ( $WarnMsg ) { Write-LogMessage -Type WARNING -Message $WarnMsg -Colour Magenta } if ( $ErrorMsg ) { Write-LogMessage -Type ERROR -Message $ErrorMsg -Colour Red }
 }